⚡ Improves Startup Time for Noninteractive Session

+++ powershell/hooks/rc/_ls.ps1
@@ -3,5 +3,9 @@ $ErrorActionPreference = "Stop"
 Set-StrictMode -Version Latest
 
 
+if (-not (Test-SessionInteractivity)) {
+    return
+}
+
 Remove-Alias -Name sl -Force
 phook_enqueue_module "poshy-wrap-ls"
